+++ prometheus/test_locally.sh
@@ -3,29 +3,31 @@
 
 export GOSS_WAIT_OPTS="-r 90s -s 1s > /dev/null"
 
+PROMETHEUS_VERSION='v2.8.0'
+GRAFANA_VERSION='6.0.2'
+
+make
+
 docker network create nodejs_prom
 # docker build . -t jecnua/nodejs-monitor-testapp
 
-docker stop nodejs-monitor-testapp
-docker rm nodejs-monitor-testapp
+docker rm -f nodejs-monitor-testapp
 docker run \
   --name nodejs-monitor-testapp \
   --network nodejs_prom \
   -d \
   -p 3001:3001 \
-  jecnua/nodejs-monitor-testapp:latest
+  jecnua/monitor-nodejs-prom:dev-latest
 
-docker stop prometheus
-docker rm prometheus
+docker rm -f prometheus
 docker run --name prometheus \
   --network nodejs_prom \
   -v "$PWD"/prometheus.yml:/etc/prometheus/prometheus.yml:ro \
   -d \
   -p 9090:9090 \
-  quay.io/prometheus/prometheus:v2.2.1
+  "quay.io/prometheus/prometheus:$PROMETHEUS_VERSION"
 
-docker stop grafana
-docker rm grafana
+docker rm -f grafana
 docker run -d --name=grafana \
   -e 'GF_INSTALL_PLUGINS=grafana-clock-panel,grafana-piechart-panel' \
   -e 'GF_AUTH_ANONYMOUS_ENABLED=true' \
@@ -36,4 +38,4 @@ docker run -d --name=grafana \
   -v "$PWD"/nodejs.json:/var/lib/grafana/dashboards/nodejs.json:ro \
   -v "$PWD"/prom.yml:/etc/grafana/provisioning/datasources/datasource.yaml:ro \
   -p 3000:3000 \
-  grafana/grafana:5.1.0
+  "grafana/grafana:$GRAFANA_VERSION"