+++ prometheus/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.0
 
 - Nodejs app working correctly with prometheus exporter.
+- Updated grafana and prometheus version
 
 ### BREAKING CHANGES